# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LegalGuard AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuditEvent {
  id: number;
  timestamp: string;
  user: string;
  agent: string;
  action: string;
  status: string;
  hash: string;
  /** @nullable */
  receiptId?: string | null;
  /** @nullable */
  contractId?: number | null;
  /** @nullable */
  contractName?: string | null;
  /** @nullable */
  details?: string | null;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface RiskCount {
  level: string;
  count: number;
}

export interface DashboardSummary {
  pendingApprovals: number;
  contractsProcessed: number;
  blockedActions: number;
  aiConfidence: number;
  averageRiskScore: number;
  recentActivity: AuditEvent[];
  agentHealth: string;
  contractsByStatus: StatusCount[];
  approvalsByRisk: RiskCount[];
}

export interface Contract {
  id: number;
  name: string;
  client: string;
  status: string;
  riskScore: number;
  riskLevel: string;
  uploadedAt: string;
  /** @nullable */
  analyzedAt?: string | null;
  fileType: string;
  fileSize: string;
  /** @nullable */
  assignedTo?: string | null;
}

export interface Clause {
  id: number;
  title: string;
  content: string;
  riskLevel: string;
  isRisky: boolean;
  /** @nullable */
  aiSuggestion?: string | null;
}

export interface ContractDetail {
  id: number;
  name: string;
  client: string;
  status: string;
  riskScore: number;
  riskLevel: string;
  uploadedAt: string;
  /** @nullable */
  analyzedAt?: string | null;
  fileType: string;
  fileSize: string;
  /** @nullable */
  assignedTo?: string | null;
  /** @nullable */
  summary?: string | null;
  clauses: Clause[];
  obligations: string[];
  deadlines: string[];
}

export interface ContractAnalysis {
  contractId: number;
  overallRiskScore: number;
  riskLevel: string;
  executiveSummary: string;
  riskyClauses: Clause[];
  missingClauses: string[];
  aiSuggestions: string[];
  confidenceScore: number;
  sensitiveActions: string[];
}

export interface ContractInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  client: string;
  fileType: string;
  fileSize: string;
  /** @nullable */
  content?: string | null;
}

export interface ContractUpdate {
  name?: string;
  client?: string;
  status?: string;
  /** @nullable */
  assignedTo?: string | null;
}

export interface Approval {
  id: number;
  contractId: number;
  contractName: string;
  client: string;
  requestedAction: string;
  requestedBy: string;
  reason: string;
  aiExplanation: string;
  riskScore: number;
  riskLevel: string;
  requiredApprover: string;
  estimatedImpact: string;
  status: string;
  createdAt: string;
  /** @nullable */
  resolvedAt?: string | null;
  /** @nullable */
  resolvedBy?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface ApprovalDecision {
  /** @nullable */
  notes?: string | null;
}

export interface CryptoReceipt {
  id: number;
  receiptId: string;
  transactionHash: string;
  documentName: string;
  requestedBy: string;
  approvedBy: string;
  timestamp: string;
  signature: string;
  verificationStatus: string;
  /** @nullable */
  approvalId?: number | null;
}

export interface Notification {
  id: number;
  title: string;
  description: string;
  category: string;
  status: string;
  createdAt: string;
  isRead: boolean;
  /** @nullable */
  cta?: string | null;
  /** @nullable */
  ctaLink?: string | null;
}

export interface TeamMember {
  id: number;
  name: string;
  email: string;
  role: string;
  status: string;
  joinedAt: string;
  /** @nullable */
  lastActive?: string | null;
  /** @nullable */
  avatar?: string | null;
  actionsCount: number;
  approvalsCount: number;
}

export interface TeamMemberInvite {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  email: string;
  /** @minLength 1 */
  role: string;
}

export interface ChatMessage {
  id: number;
  role: string;
  content: string;
  timestamp: string;
  /** @nullable */
  contractId?: number | null;
  isBlocked: boolean;
  /** @nullable */
  blockedAction?: string | null;
  /** @nullable */
  approvalId?: number | null;
  /** @nullable */
  confidenceScore?: number | null;
  evidenceRefs: string[];
}

export interface ChatMessageInput {
  /** @minLength 1 */
  message: string;
  /** @nullable */
  contractId?: number | null;
}

export type ListContractsParams = {
status?: string;
search?: string;
};

export type ListApprovalsParams = {
status?: string;
};

export type ListAuditEventsParams = {
search?: string;
status?: string;
from?: string;
to?: string;
};

export type ListNotificationsParams = {
category?: string;
};

export type GetChatHistoryParams = {
/**
 * @nullable
 */
contractId?: number | null;
};