# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LegalGuard AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /approvals:
    get:
      operationId: listApprovals
      tags: [approvals]
      summary: List all approval requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of approvals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Approval"
  /approvals/{id}:
    get:
      operationId: getApproval
      tags: [approvals]
      summary: Get approval details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approval details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
  /approvals/{id}/approve:
    post:
      operationId: approveAction
      tags: [approvals]
      summary: Approve an action
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecision"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
  /approvals/{id}/reject:
    post:
      operationId: rejectAction
      tags: [approvals]
      summary: Reject an action
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecision"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
...
  /assistant/chat:
    post:
      operationId: sendChatMessage
      tags: [assistant]
      summary: Send a message to the AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
  /assistant/history:
    get:
      operationId: getChatHistory
      tags: [assistant]
      summary: Get chat history
      parameters:
        - name: contractId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
components:
  schemas:
...
    Approval:
      type: object
      properties:
        id:
          type: integer
        contractId:
          type: integer
        contractName:
          type: string
        client:
          type: string
        requestedAction:
          type: string
        requestedBy:
          type: string
        reason:
          type: string
        aiExplanation:
          type: string
        riskScore:
          type: number
        riskLevel:
          type: string
        requiredApprover:
          type: string
        estimatedImpact:
          type: string
        status:
          type: string
        createdAt:
          type: string
        resolvedAt:
          type: ["string", "null"]
        resolvedBy:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [id, contractId, contractName, client, requestedAction, requestedBy, reason, aiExplanation, riskScore, riskLevel, requiredApprover, estimatedImpact, status, createdAt]
    ApprovalDecision:
      type: object
      properties:
        notes:
          type: ["string", "null"]
      required: []
...
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        role:
          type: string
        content:
          type: string
        timestamp:
          type: string
        contractId:
          type: ["integer", "null"]
        isBlocked:
          type: boolean
        blockedAction:
          type: ["string", "null"]
        approvalId:
          type: ["integer", "null"]
        confidenceScore:
          type: ["number", "null"]
        evidenceRefs:
          type: array
          items:
            type: string
      required: [id, role, content, timestamp, isBlocked, evidenceRefs]
    ChatMessageInput:
      type: object
      properties:
        message:
          type: string
          minLength: 1
        contractId:
          type: ["integer", "null"]
      required: [message]