# lib/api-zod/src/generated/types/approval.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LegalGuard AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Approval {
  id: number;
  contractId: number;
  contractName: string;
  client: string;
  requestedAction: string;
  requestedBy: string;
  reason: string;
  aiExplanation: string;
  riskScore: number;
  riskLevel: string;
  requiredApprover: string;
  estimatedImpact: string;
  status: string;
  createdAt: string;
  /** @nullable */
  resolvedAt?: string | null;
  /** @nullable */
  resolvedBy?: string | null;
  /** @nullable */
  notes?: string | null;
}